+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - **Strings**
   - Add the `chunk()` function : Splits a string into groups of length, separated by a separator.
   - Add the `replace()` function : Replaces all occurrences of a substring within a string, with optional Unicode (grapheme-safe) and normalization support.
+  - Add the `split()` function : Splits a string into an array using a regular expression separator.
   
 ## [1.0.7] - 2025-12-12